requirements.txt
streamlit
pandas
supabase
yfinance
scipy
openpyxl
python-dotenv
plotly